Split boostrap in _tasks_ (for tests/coverage) and _cli_ (to run with `npx`)
